removed useless files, added finished files
